Proof of concept using Ignite React Native starter
This is completely different than the master branch. Checkout this
branch after fetching and then read the README.md file on using. You can
then get started by looking at the app when you run iOS version and see
some changes I started. what I like about this is that is gives some
great ideas to work off of, even if there are only parts of it that are
used in the app.